us-news/requirements.txt
boto3==1.35.99
python-dotenv==1.0.1 